requirements.txt
pandas
forexconnect
datetime
scipy
python-dotenv